refactor(test-utils): relocate renderPortalWithProviders inside renderer.tsx
