+++ packages/react/src/components/modal/modal-dialog.test.tsx
@@ -1,10 +1,9 @@
 import { fireEvent, RenderResult } from '@testing-library/react';
 import { doNothing } from '../../test-utils/callbacks';
-import { renderPortalWithProviders } from '../../test-utils/portal-renderer';
 import { DeviceType } from '../device-context-provider/device-context-provider';
 import { ModalDialog, ModalDialogProps } from './modal-dialog';
 import { getByTestId as enzymeGetByTestId } from '../../test-utils/enzyme-selectors';
-import { mountWithProviders } from '../../test-utils/renderer';
+import { mountWithProviders, renderPortalWithProviders } from '../../test-utils/renderer';
 
 jest.mock('../../utils/uuid');
