+++ packages/react/src/components/modal/modal.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, RenderResult } from '@testing-library/react';
 import { doNothing } from '../../test-utils/callbacks';
-import { renderPortalWithProviders } from '../../test-utils/portal-renderer';
+import { renderPortalWithProviders } from '../../test-utils/renderer';
 import { DeviceType } from '../device-context-provider/device-context-provider';
 import { Modal, ModalProps } from './modal';
