+++ packages/react/src/test-utils/renderer.tsx
@@ -3,6 +3,7 @@ import { Component, FunctionComponent, ReactElement, ReactPortal } from 'react';
 import ReactDOM from 'react-dom';
 import { act } from 'react-dom/test-utils';
 import { MemoryRouter } from 'react-router-dom';
+import { render as testingLibRender, RenderResult } from '@testing-library/react';
 import { DesignSystem, DesignSystemProps } from '../components/design-system';
 import { DeviceType } from '../components/device-context-provider/device-context-provider';
 import { ThemeWrapper } from '../components/theme-wrapper/theme-wrapper';
@@ -76,4 +77,11 @@ export async function actAndWaitForEffects<C extends Component, P = C['props'], 
     await act(async () => action());
     await actUpdate(wrapper);
     return wrapper;
+}
+
+export function renderPortalWithProviders(
+    component: ReactElement,
+    device?: DeviceType,
+): RenderResult {
+    return testingLibRender(<AllProviders staticDevice={device}>{component}</AllProviders>);
 }
